requirements.txt
colorama==0.4.6
PyYAML==6.0.1
requests==2.31.0
pygame>=2.1.0
